backend1/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
PyJWT==2.8.0
